Feat: Migrate to module, and docker build
* Add dockerfile for a easy to use basic deployment
* Migrate to a python module (I wanted to pip install in my docker image)
* Setup pyproject.toml (pip install $url)
* Migrate dotenv to the app.py wrapper
* beetiful/__init__.py is just app.py but git thinks its a new file
