--- requirements.txt
@@ -1,19 +1,0 @@
-beautifulsoup4==4.12.3
-blinker==1.8.2
-bs4==0.0.2
-certifi==2024.8.30
-charset-normalizer==3.3.2
-click==8.1.7
-flask==3.0.3
-idna==3.9
-importlib-metadata==8.5.0
-itsdangerous==2.2.0
-jinja2==3.1.4
-MarkupSafe==2.1.5
-python-dotenv==1.0.1
-PyYAML==6.0.2
-requests==2.32.3
-soupsieve==2.6
-urllib3==2.2.3
-werkzeug==3.0.4
-zipp==3.20.2
